Digital-Service/website/requirements.txt
pandas 
mysql-connector-python

#pip install -r requirements.txt